base: Add -debian10 for all images; include in all to publish
Experimental Debian 10 support was added in commit ee7342d305, but was
missing -debian10 variants for some of the images in base. This adds
these variants for all the images in the //base package. It also adds
them to the //:all target, which might mean they get published.

base.bzl: Use a function to define the repeated images/tests. This
  should be slightly easier to understand than a list comprehension.
testdata/debian(9|10).yaml: Tests for the os-release package. This
  caught at least one bug where I missed the distro_suffix.

package_manager: Add support for Debian 10 buster. This will need an
  update to package_manager.bzl once the change is published.