cloudbuild.yaml: Publish -debian9 and -debian10 images
This supports the base, cc, and python images.

This keeps the existing containers, but adds -debian9 and -debian10
variants. I wrote a script to generate cloudbuild.yaml, which I can
add, or will submit if this change is merged, since managing this
manually is now challenging.

This depends on my changes to add the -debian9 suffix.

I tested this by publishing these images to my own project: