Mention how to append the debug tag when an image has a tag already
Hi Distroless, 

Thanks for documenting the `:debug` option to allow troubleshooting with a shell, very helpful. It wasn't very obvious to me from the README.me how to apply the debug tag to an image that already has a tag, like `gcr.io/distroless/java:11`. 

I thought adding a dash would help and it worked, but I'm not sure if this is the recommended way to do this. Opening this PR to start some discussion.